Cast update due to the use of G++ flag -Wconversion and -Wfatal-errors
git-svn-id: http://130.66.124.6/svn/maroff/Product/Cpp/Tools/interpolation@1706 300066c3-8af3-463d-8cc0-c0233cc64ac1

+++ interpolation/src/LinearInterpolation.cpp
@@ -24,7 +24,7 @@ void LinearInterpolation::compute_coefficients_for_ith_interval(const double val
 {
     (void) i;
     (void) val;
-    x0[idx] = xmin + idx*delta;
+    x0[idx] = xmin + double(idx)*delta;
     x1[idx] = x0[idx] + delta;
     y0[idx] = y.at(idx);
     y1[idx] = (n==1) ? y0[idx] : y.at(idx+1);
